'修改ID数据类型'

+++ rms-common/src/main/java/com/still/rms/common/dto/ResourceDto.java
@@ -17,5 +17,5 @@ import java.util.List;
 public class ResourceDto extends Resource {
 
     @ApiModelProperty(value = "演员ID数组")
-    private List<Integer> actors;
+    private List<Long> actors;
 }
